requirements.txt
flask==2.3.3
flask-cors==4.0.0
supabase==2.0.2
qrcode==7.4.2
pillow==10.0.1
werkzeug==2.3.7
python-dotenv==1.0.0

